+++ front-end/package.json
@@ -30,5 +30,10 @@
     "eslint-plugin-react-refresh": "^0.4.18",
     "globals": "^15.14.0",
     "vite": "^6.1.0"
-  }
+  },
+  "description": "- to install run:\r   ```npm install```\r - to run the server:\r   ```npm run dev```",
+  "main": "eslint.config.js",
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
 }
