+++ front-end/src/App/App.jsx
@@ -1,118 +1,9 @@
 import { useState, useEffect } from "react";
 import "./App.css";
-
-
-function AboutExercise({ onClose }) {
-  return (
-    <div className="about">
-      <div className="about-content">
-        <h2>About Us!</h2>
-        <p>
-          <strong>Creators of the website:</strong>
-        </p>
-        <ul>
-          <li>Tomas Moshi: 3rd year Computer Science and Engineering major</li>
-          <li>Adhia Amini: 3rd year Computer Science and Engineering major</li>
-          <li>Sina Ghadimi: 4th year Electrical Engineering major</li>
-          <li>Amaras Issagholian: 3rd year Computer Science major</li>
-          <li>Shervin Shahidi Asl: 3rd year Computer Science major</li>
-        </ul>
-        <p>
-          As students we understand that it is difficult to maintain a healthy 
-          balance between academics and having a healthy life style. We built this platform 
-          in order to help students find easy and fun exercises like tennis, basketball, and yoga. 
-          We aim to connect students with similar interests in order to build long lasting relationships.
-          
-        </p>
-        <button onClick={onClose}>Close</button>
-      </div>
-    </div>
-  );
-}
-
-
-function AboutUsButton() {
-  const [showAbout, setShowAbout] = useState(false);
-
-  const handleButtonClick = () => {
-    setShowAbout(true);
-  };
-
-  const handleCloseAbout = () => {
-    setShowAbout(false);
-  };
-
-  return (
-    <div>
-      <button onClick={handleButtonClick} className="about-us-button">
-        About Us
-      </button>
-      {showAbout && <AboutExercise onClose={handleCloseAbout} />}
-    </div>
-  );
-}
-
-
-function HelpModal({ onClose }) {
-  return (
-    <div className="modal">
-      <div className="modal-content">
-        <h2>Welcome to AllExercises!</h2>
-        <p>Here are some instructions to help you get started with AllExercises:</p>
-        <ul>
-          <li>
-            <strong>Sign Up:</strong> Click the "Sign up" button at the top right and create an account. Verify via email and you're set!
-          </li>
-          <li>
-            <strong>After logging in</strong> you can search for different sports/workouts and even post events for others to join.
-          </li>
-          <li>
-            <strong>Post an Event:</strong>
-            <ul>
-              <li>Navigate to the "Post Event" section.</li>
-              <li>Select an exercise type (e.g., Basketball, Tennis, etc.).</li>
-              <li>Fill in the details (location, date, time, etc.).</li>
-              <li>Submit your event for others to join.</li>
-            </ul>
-          </li>
-          <li>
-            <strong>Join an Event:</strong>
-            <ul>
-              <li>Browse the list of upcoming events on our homepage.</li>
-              <li>Click on an event to view details.</li>
-              <li>If interested, click the "Join" button to participate.</li>
-            </ul>
-          </li>
-        </ul>
-        <button onClick={onClose}>Close</button>
-      </div>
-    </div>
-  );
-}
-
-
-function NeedHelpButton() {
-  const [showModal, setShowModal] = useState(false);
-
-  const handleButtonClick = () => {
-    setShowModal(true);
-  };
-
-  const handleCloseModal = () => {
-    setShowModal(false);
-  };
-
-  return (
-    <div>
-      <button onClick={handleButtonClick} className="need-help-button">
-        Need Help?
-      </button>
-      {showModal && <HelpModal onClose={handleCloseModal} />}
-    </div>
-  );
-}
-
-
+import Login from "../Login.jsx"; // Ensure this path matches the actual file name
+import Home from "../Features/AboutUs/AboutUs.jsx";
+import Home from "../Features/AboutUs/AboutUsButton.jsx";
+import Home from "../Features/AboutUs/HelpModal.jsx";
 function App() {
   const [searchQuery, setSearchQuery] = useState("");
   const [isLoginOpen, setIsLoginOpen] = useState(false);
@@ -179,8 +70,7 @@ function App() {
           <li><strong>Pilates</strong></li>
         </ul>
       </section>
-
-      {}
+      {isLoginOpen && <Login isSignUpMode={isSignUpMode} onClose={closeLogin} />}
       <NeedHelpButton />
       <AboutUsButton />
     </div>
